refactor tests for fa to regex transitions
