+++ pyformlang/finite_automaton/tests/test_epsilon.py
@@ -1,6 +1,7 @@
 """
 Tests for epsilon transitions
 """
+
 from pyformlang.finite_automaton import Epsilon
 from pyformlang.finite_automaton import Symbol
