+++ pyformlang/finite_automaton/tests/test_state.py
@@ -1,6 +1,7 @@
 """
 Tests the states
 """
+
 from pyformlang.finite_automaton import State
