+++ pyformlang/regular_expression/tests/test_python_regex.py
@@ -1,6 +1,7 @@
 """
 Testing python regex parsing
 """
+
 import re
 
 from pyformlang.regular_expression.python_regex import PythonRegex
