+++ pyformlang/regular_expression/tests/test_regex.py
@@ -1,15 +1,22 @@
 """
 Tests for regular expressions
 """
-from pyformlang.regular_expression import Regex, MisformedRegexError
-from pyformlang import finite_automaton
+
 import pytest
+
+from pyformlang.regular_expression import Regex, MisformedRegexError
+from pyformlang.finite_automaton import EpsilonNFA
+from pyformlang.finite_automaton import DeterministicFiniteAutomaton
+from pyformlang.finite_automaton import State, Symbol, Epsilon
+from pyformlang.finite_automaton.tests.test_deterministic_finite_automaton \
+    import get_example0, get_dfa_example, perform_tests_example0
 
 
 class TestRegex:
     """ Tests for regex """
 
     # pylint: disable=missing-function-docstring,too-many-public-methods
+    # pylint: disable=protected-access
 
     def test_creation(self):
         """ Try to create regex """
@@ -65,10 +72,10 @@ class TestRegex:
 
     def test_to_enfa0(self):
         """ Tests the transformation to a regex """
-        symb_a = finite_automaton.Symbol("a")
-        symb_b = finite_automaton.Symbol("b")
-        symb_c = finite_automaton.Symbol("c")
-        epsilon = finite_automaton.Epsilon()
+        symb_a = Symbol("a")
+        symb_b = Symbol("b")
+        symb_c = Symbol("c")
+        epsilon = Epsilon()
         regex = Regex("a|b")
         enfa = regex.to_epsilon_nfa()
         assert enfa.accepts([symb_a])
@@ -107,9 +114,9 @@ class TestRegex:
 
     def test_to_enfa1(self):
         """ Tests the transformation to a regex """
-        symb_a = finite_automaton.Symbol("a")
-        symb_b = finite_automaton.Symbol("b")
-        symb_c = finite_automaton.Symbol("c")
+        symb_a = Symbol("a")
+        symb_b = Symbol("b")
+        symb_c = Symbol("c")
         regex = Regex("a**")
         enfa = regex.to_epsilon_nfa()
         assert enfa.accepts([symb_a])
@@ -279,3 +286,193 @@ class TestRegex:
     def test_backslash(self):
         assert Regex("(\\\\|])").accepts("\\")
         assert Regex("(\\\\|])").accepts("]")
+
+    def test_remove_state(self):
+        " Tests the remove of state """
+        enfa = EpsilonNFA()
+        state0 = State(0)
+        state1 = State(1)
+        state2 = State(2)
+        symb02 = Symbol("a+b")
+        symb01 = Symbol("c*")
+        symb11 = Symbol("b+(c.d)")
+        symb12 = Symbol("a.b.c")
+        enfa.add_start_state(state0)
+        enfa.add_final_state(state2)
+        enfa.add_transition(state0, symb01, state1)
+        enfa.add_transition(state0, symb02, state2)
+        enfa.add_transition(state1, symb11, state1)
+        enfa.add_transition(state1, symb12, state2)
+        Regex._remove_all_basic_states(enfa)
+        assert enfa.get_number_transitions() == 1
+        assert len(enfa.states) == 2
+
+    def test_from_enfa1(self):
+        """ Tests the transformation to regex """
+        enfa = EpsilonNFA()
+        state0 = State(0)
+        state1 = State(1)
+        state2 = State(2)
+        symb_e = Symbol("e")
+        symb_f = Symbol("f")
+        symb_g = Symbol("g")
+        enfa.add_start_state(state0)
+        enfa.add_final_state(state2)
+        enfa.add_transition(state0, symb_e, state1)
+        enfa.add_transition(state1, symb_f, state2)
+        enfa.add_transition(state0, symb_g, state2)
+        regex = Regex.from_finite_automaton(enfa)
+        enfa2 = regex.to_epsilon_nfa()
+        assert enfa2.accepts([symb_e, symb_f])
+        assert enfa2.accepts([symb_g])
+        assert not enfa2.accepts([])
+        assert not enfa2.accepts([symb_e])
+        assert not enfa2.accepts([symb_f])
+        enfa.add_final_state(state0)
+        with pytest.raises(ValueError) as _:
+            Regex._get_regex_simple(enfa)
+        regex = Regex.from_finite_automaton(enfa)
+        enfa3 = regex.to_epsilon_nfa()
+        assert enfa3.accepts([symb_e, symb_f])
+        assert enfa3.accepts([symb_g])
+        assert enfa3.accepts([])
+        assert not enfa3.accepts([symb_e])
+        assert not enfa3.accepts([symb_f])
+        enfa.remove_start_state(state0)
+        regex = Regex.from_finite_automaton(enfa)
+        enfa3 = regex.to_epsilon_nfa()
+        assert not enfa3.accepts([symb_e, symb_f])
+        assert not enfa3.accepts([symb_g])
+        assert not enfa3.accepts([])
+        assert not enfa3.accepts([symb_e])
+        assert not enfa3.accepts([symb_f])
+        enfa.add_start_state(state0)
+        enfa.add_transition(state0, symb_f, state0)
+        regex = Regex.from_finite_automaton(enfa)
+        enfa3 = regex.to_epsilon_nfa()
+        assert enfa3.accepts([symb_e, symb_f])
+        assert enfa3.accepts([symb_f, symb_e, symb_f])
+        assert enfa3.accepts([symb_g])
+        assert enfa3.accepts([symb_f, symb_f, symb_g])
+        assert enfa3.accepts([])
+        assert not enfa3.accepts([symb_e])
+        assert enfa3.accepts([symb_f])
+
+    def test_from_enfa2(self):
+        """ Tests the transformation to regex """
+        enfa = EpsilonNFA()
+        state0 = State(0)
+        state1 = State(1)
+        symb_a = Symbol("0")
+        symb_b = Symbol("1")
+        enfa.add_start_state(state0)
+        enfa.add_final_state(state1)
+        enfa.add_transition(state0, symb_a, state0)
+        enfa.add_transition(state0, symb_a, state1)
+        enfa.add_transition(state1, symb_b, state0)
+        enfa.add_transition(state1, symb_b, state1)
+        regex = Regex.from_finite_automaton(enfa)
+        enfa2 = regex.to_epsilon_nfa()
+        assert enfa2.accepts([symb_a])
+        assert enfa2.accepts([symb_a, symb_a])
+        assert enfa2.accepts([symb_a, symb_a, symb_b])
+        assert enfa2.accepts([symb_a, symb_a, symb_b, symb_b])
+        assert enfa2.accepts([symb_a, symb_a,
+                                       symb_b, symb_b, symb_a])
+        assert enfa2.accepts([symb_a, symb_a, symb_b,
+                                       symb_b, symb_a, symb_b])
+        assert not enfa2.accepts([symb_b])
+
+    def test_from_enfa3(self):
+        """ Tests the transformation to regex """
+        enfa = EpsilonNFA()
+        state0 = State(0)
+        state1 = State(1)
+        symb_a = Symbol("0")
+        symb_b = Symbol("1")
+        enfa.add_start_state(state0)
+        enfa.add_final_state(state1)
+        enfa.add_transition(state0, symb_a, state0)
+        enfa.add_transition(state1, symb_b, state0)
+        enfa.add_transition(state1, symb_b, state1)
+        regex = Regex.from_finite_automaton(enfa)
+        enfa2 = regex.to_epsilon_nfa()
+        assert not enfa2.accepts([symb_a])
+        assert not enfa2.accepts([symb_a, symb_a])
+        assert not enfa2.accepts([symb_a, symb_a, symb_b])
+        assert not enfa2.accepts([symb_a, symb_a,
+                                        symb_b, symb_b, symb_a])
+        assert not enfa2.accepts([symb_a, symb_a, symb_b,
+                                        symb_b, symb_a, symb_b])
+        assert not enfa2.accepts([symb_b])
+        epsilon = Epsilon()
+        enfa.add_transition(state0, epsilon, state1)
+        regex = Regex.from_finite_automaton(enfa)
+        enfa2 = regex.to_epsilon_nfa()
+        assert enfa.accepts([])
+        assert enfa.accepts([symb_a])
+        assert enfa2.accepts([symb_a])
+        assert enfa2.accepts([symb_a, symb_a])
+        assert enfa2.accepts([symb_a, symb_a, symb_b, symb_b])
+        assert enfa2.accepts([symb_a, symb_a, symb_b, symb_b,
+                                       symb_a, symb_b])
+        assert enfa2.accepts([symb_b])
+        assert enfa2.accepts([])
+        enfa.remove_transition(state0, symb_a, state0)
+        regex = Regex.from_finite_automaton(enfa)
+        enfa2 = regex.to_epsilon_nfa()
+        assert not enfa2.accepts([symb_a])
+        assert not enfa2.accepts([symb_a, symb_a])
+        assert not enfa2.accepts([symb_a, symb_a, symb_b])
+        assert not enfa2.accepts([symb_a, symb_a, symb_b,
+                                        symb_b, symb_a])
+        assert not enfa2.accepts([symb_a, symb_a, symb_b, symb_b,
+                                        symb_a, symb_b])
+        assert enfa2.accepts([symb_b])
+        assert enfa2.accepts([])
+        enfa.remove_transition(state1, symb_b, state1)
+        regex = Regex.from_finite_automaton(enfa)
+        enfa2 = regex.to_epsilon_nfa()
+        assert enfa2.accepts([symb_b, symb_b])
+        enfa.add_transition(state0, symb_a, state0)
+        regex = Regex.from_finite_automaton(enfa)
+        enfa2 = regex.to_epsilon_nfa()
+        assert enfa2.accepts([symb_a, symb_b])
+
+    def test_example_doc(self):
+        enfa0 = EpsilonNFA()
+        state0 = State(0)
+        state1 = State(1)
+        symb_a = Symbol("0")
+        symb_b = Symbol("1")
+        enfa0.add_start_state(state0)
+        enfa0.add_final_state(state1)
+        enfa0.add_transition(state0, symb_a, state0)
+        enfa0.add_transition(state1, symb_b, state0)
+        enfa0.add_transition(state1, symb_b, state1)
+
+        # Turn a finite automaton into a regex...
+        regex = Regex.from_finite_automaton(enfa0)
+        # And turn it back into an epsilon non deterministic automaton
+        enfa1 = regex.to_epsilon_nfa()
+        dfa0 = DeterministicFiniteAutomaton.from_epsilon_nfa(enfa0)
+        dfa1 = DeterministicFiniteAutomaton.from_epsilon_nfa(enfa1)
+        assert dfa0.is_equivalent_to(dfa1)
+
+    def test_from_dfa0(self):
+        """ Tests the regex transformation """
+        dfa0 = get_example0()
+        enfa = Regex.from_finite_automaton(dfa0).to_epsilon_nfa()
+        perform_tests_example0(enfa)
+
+    def test_from_dfa1(self):
+        dfa1 = get_dfa_example()
+        enfa = Regex.from_finite_automaton(dfa1).to_epsilon_nfa()
+        dfa2 = DeterministicFiniteAutomaton.from_epsilon_nfa(enfa)
+        assert dfa1.is_equivalent_to(dfa2)
+
+    def test_to_minimal_dfa(self):
+        dfa0 = get_example0()
+        dfa_regex = Regex.from_finite_automaton(dfa0)
+        dfa1 = dfa_regex.to_minimal_dfa()
+        assert dfa0 == dfa1
